refactor(app): move render view setup out of event loop
